+++ Solvers/Wizards/Leandro/Leandro01.cs
@@ -12,7 +12,7 @@ namespace Solvers
         {
         }
 
-        public override int SolvePartOne(string[] input)
+        public override long SolvePartOne(string[] input)
         {           
             int[] numInput = new int[input.Count()];
             int solution = 0;
@@ -50,7 +50,7 @@ namespace Solvers
             return solution;
         }
 
-        public override int SolvePartTwo(string[] input)
+        public override long SolvePartTwo(string[] input)
         {
             int[] numInput = new int[input.Count()];
             int solution = 0;
